Delete old AudioPlayer and replace it with AudioPlayerNew
