+++ cardigan/stories/components/AudioPlayer/AudioPlayer.stories.tsx
@@ -1,8 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { ReadmeDecorator } from '@weco/cardigan/config/decorators';
+import { prismicRichTextMultiline } from '@weco/cardigan/stories/data/text';
 import AudioPlayer from '@weco/content/components/AudioPlayer';
-import Readme from '@weco/content/components/AudioPlayer/README.mdx';
 
 const meta: Meta<typeof AudioPlayer> = {
   title: 'Components/AudioPlayer',
@@ -11,6 +10,20 @@ const meta: Meta<typeof AudioPlayer> = {
     audioFile:
       'https://iiif.wellcomecollection.org/av/b2248887x_0001.wav/full/max/default.mp3#identity',
     title: 'Mat Fraser: interview 1',
+    isDark: true,
+    transcript: prismicRichTextMultiline,
+  },
+  argTypes: {
+    titleProps: {
+      table: {
+        disable: true,
+      },
+    },
+    transcript: {
+      table: {
+        disable: true,
+      },
+    },
   },
 };
 
@@ -21,10 +34,10 @@ type Story = StoryObj<typeof AudioPlayer>;
 export const Basic: Story = {
   name: 'AudioPlayer',
   render: args => (
-    <ReadmeDecorator
-      WrappedComponent={AudioPlayer}
-      args={args}
-      Readme={Readme}
-    />
+    <div
+      style={{ padding: '20px', background: args.isDark ? '#111' : 'white' }}
+    >
+      <AudioPlayer {...args} />
+    </div>
   ),
 };
