--- content/webapp/components/AudioPlayer/README.md
@@ -1,5 +1,0 @@
-## Purpose
-Provides a simple accessible custom UI for an HTML `<audio>` element.
-
-
-[Edit this Readme on GitHub](https://github.com/wellcomecollection/wellcomecollection.org/edit/main/common/views/components/AudioPlayer/README.md)
